Added Form Switching
Added Automatically switching forms if the conditions are met:

Conditions are:

Human:
If in human and javelin toss is on cooldown and the target is lower the
pounce range it will switch forms and perform the cougar combo.

Cougar:
If in cougar form and pounce isn't ready to be casted and also if the
distance between the player and the target is higher then the total
pounce distance and if the javelin toss has a custom hitchance option of
hitting the target, it will switch forms.